Moved StaticStore to Bit.Core.Billing.Models.StaticStore namespace
