+++ src/Core/Billing/Models/StaticStore/Plans/EnterprisePlan2023.cs
@@ -1,7 +1,7 @@
 ﻿using Bit.Core.Billing.Enums;
-using Bit.Core.Enums;
+using Bit.Core.Models.StaticStore;
 
-namespace Bit.Core.Models.StaticStore.Plans;
+namespace Bit.Core.Billing.Models.StaticStore.Plans;
 
 public record Enterprise2023Plan : Plan
 {
