+++ src/Core/Billing/Models/StaticStore/Plans/Families2019Plan.cs
@@ -1,9 +1,9 @@
 ﻿using Bit.Core.Billing.Enums;
-using Bit.Core.Enums;
+using Bit.Core.Models.StaticStore;
 
-namespace Bit.Core.Models.StaticStore.Plans;
+namespace Bit.Core.Billing.Models.StaticStore.Plans;
 
-public record Families2019Plan : Models.StaticStore.Plan
+public record Families2019Plan : Plan
 {
     public Families2019Plan()
     {
