+++ src/Core/Billing/Models/StaticStore/Plans/FreePlan.cs
@@ -1,9 +1,9 @@
 ﻿using Bit.Core.Billing.Enums;
-using Bit.Core.Enums;
+using Bit.Core.Models.StaticStore;
 
-namespace Bit.Core.Models.StaticStore.Plans;
+namespace Bit.Core.Billing.Models.StaticStore.Plans;
 
-public record FreePlan : Models.StaticStore.Plan
+public record FreePlan : Plan
 {
     public FreePlan()
     {
