+++ src/Core/Billing/Models/StaticStore/Plans/Teams2019Plan.cs
@@ -1,15 +1,15 @@
 ﻿using Bit.Core.Billing.Enums;
-using Bit.Core.Enums;
+using Bit.Core.Models.StaticStore;
 
-namespace Bit.Core.Models.StaticStore.Plans;
+namespace Bit.Core.Billing.Models.StaticStore.Plans;
 
-public record TeamsPlan : Plan
+public record Teams2019Plan : Plan
 {
-    public TeamsPlan(bool isAnnual)
+    public Teams2019Plan(bool isAnnual)
     {
-        Type = isAnnual ? PlanType.TeamsAnnually : PlanType.TeamsMonthly;
+        Type = isAnnual ? PlanType.TeamsAnnually2019 : PlanType.TeamsMonthly2019;
         ProductTier = ProductTierType.Teams;
-        Name = isAnnual ? "Teams (Annually)" : "Teams (Monthly)";
+        Name = isAnnual ? "Teams (Annually) 2019" : "Teams (Monthly) 2019";
         IsAnnual = isAnnual;
         NameLocalizationKey = "planNameTeams";
         DescriptionLocalizationKey = "planDescTeams";
@@ -27,18 +27,19 @@ public record TeamsPlan : Plan
 
         UpgradeSortOrder = 3;
         DisplaySortOrder = 3;
+        LegacyYear = 2020;
 
-        PasswordManager = new TeamsPasswordManagerFeatures(isAnnual);
-        SecretsManager = new TeamsSecretsManagerFeatures(isAnnual);
+        SecretsManager = new Teams2019SecretsManagerFeatures(isAnnual);
+        PasswordManager = new Teams2019PasswordManagerFeatures(isAnnual);
     }
 
-    private record TeamsSecretsManagerFeatures : SecretsManagerPlanFeatures
+    private record Teams2019SecretsManagerFeatures : SecretsManagerPlanFeatures
     {
-        public TeamsSecretsManagerFeatures(bool isAnnual)
+        public Teams2019SecretsManagerFeatures(bool isAnnual)
         {
             BaseSeats = 0;
             BasePrice = 0;
-            BaseServiceAccount = 20;
+            BaseServiceAccount = 50;
 
             HasAdditionalSeatsOption = true;
             HasAdditionalServiceAccountOption = true;
@@ -49,27 +50,26 @@ public record TeamsPlan : Plan
             if (isAnnual)
             {
                 StripeSeatPlanId = "secrets-manager-teams-seat-annually";
-                StripeServiceAccountPlanId = "secrets-manager-service-account-2024-annually";
+                StripeServiceAccountPlanId = "secrets-manager-service-account-annually";
                 SeatPrice = 72;
-                AdditionalPricePerServiceAccount = 12;
+                AdditionalPricePerServiceAccount = 6;
             }
             else
             {
                 StripeSeatPlanId = "secrets-manager-teams-seat-monthly";
-                StripeServiceAccountPlanId = "secrets-manager-service-account-2024-monthly";
+                StripeServiceAccountPlanId = "secrets-manager-service-account-monthly";
                 SeatPrice = 7;
-                AdditionalPricePerServiceAccount = 1;
+                AdditionalPricePerServiceAccount = 0.5M;
             }
         }
     }
 
-    private record TeamsPasswordManagerFeatures : PasswordManagerPlanFeatures
+    private record Teams2019PasswordManagerFeatures : PasswordManagerPlanFeatures
     {
-        public TeamsPasswordManagerFeatures(bool isAnnual)
+        public Teams2019PasswordManagerFeatures(bool isAnnual)
         {
-            BaseSeats = 0;
+            BaseSeats = 5;
             BaseStorageGb = 1;
-            BasePrice = 0;
 
             HasAdditionalStorageOption = true;
             HasAdditionalSeatsOption = true;
@@ -78,16 +78,20 @@ public record TeamsPlan : Plan
 
             if (isAnnual)
             {
+                StripePlanId = "teams-org-annually";
                 StripeStoragePlanId = "storage-gb-annually";
-                StripeSeatPlanId = "2023-teams-org-seat-annually";
-                SeatPrice = 48;
+                StripeSeatPlanId = "teams-org-seat-annually";
+                SeatPrice = 24;
+                BasePrice = 60;
                 AdditionalStoragePricePerGb = 4;
             }
             else
             {
-                StripeSeatPlanId = "2023-teams-org-seat-monthly";
+                StripePlanId = "teams-org-monthly";
+                StripeSeatPlanId = "teams-org-seat-monthly";
                 StripeStoragePlanId = "storage-gb-monthly";
-                SeatPrice = 5;
+                BasePrice = 8;
+                SeatPrice = 2.5M;
                 AdditionalStoragePricePerGb = 0.5M;
             }
         }
