+++ src/Core/Billing/Models/StaticStore/Plans/TeamsStarterPlan2023.cs
@@ -1,7 +1,7 @@
 ﻿using Bit.Core.Billing.Enums;
-using Bit.Core.Enums;
+using Bit.Core.Models.StaticStore;
 
-namespace Bit.Core.Models.StaticStore.Plans;
+namespace Bit.Core.Billing.Models.StaticStore.Plans;
 
 public record TeamsStarterPlan2023 : Plan
 {
