+++ src/Core/Utilities/StaticStore.cs
@@ -1,9 +1,9 @@
 ﻿using System.Collections.Immutable;
 using Bit.Core.Billing.Enums;
+using Bit.Core.Billing.Models.StaticStore.Plans;
 using Bit.Core.Enums;
 using Bit.Core.Models.Data.Organizations.OrganizationUsers;
 using Bit.Core.Models.StaticStore;
-using Bit.Core.Models.StaticStore.Plans;
 
 namespace Bit.Core.Utilities;
